KAFKA-9390: Make serde pseudo-topics unique (#8054)
During the discussion for KIP-213, we decided to pass "pseudo-topics"
to the internal serdes we use to construct the wrapper serdes for
CombinedKey and hashing the left-hand-side value. However, during
the implementation, this strategy wasn't fully implemented, and we wound
up using the same topic name for a few different data types.

Cherry-pick of e16859dc48c679b3c7d9735438df046479b8ec4a from trunk

Reviewers: Guozhang Wang <guozhang@confluent.io>