KAKFA-9503: Fix TopologyTestDriver output order (#8065)
Migrates TopologyTestDriver processing to be closer to the same processing/ordering
semantics as KafkaStreams. This corrects the output order for recursive topologies
as reported in KAFKA-9503, and also works similarly in the case of task idling.

Cherry-pick of 998f1520f9af2dddfec9a9ac072f8dcf9d9004fd from trunk

Reviewers: Matthias J. Sax <matthias@confluent.io>